uses acl option to add ?public=1

+++ pages/static/js/djaodjin-pages.js
@@ -13,24 +13,28 @@
 
     jQuery(document).ready(function($) {
 
-        $("#theme-upload").djupload({
-            uploadUrl: djaodjinSettings.urls.pages.api_themes,
-        });
+        if( djaodjinSettings.urls.pages &&
+            djaodjinSettings.urls.pages.api_themes ){
 
-        $("#reset-theme").click(function(event) {
-            $.ajax({
-                method: "DELETE",
-                url: djaodjinSettings.urls.pages.api_themes,
-                datatype: "json",
-                contentType: "application/json; charset=utf-8",
-                success: function(resp) {
-                    showMessages([gettext("reset to default theme")], "success");
-                },
-                error: function(resp) {
-                    showErrorMessages(resp);
-                }
+            $("#theme-upload").djupload({
+                uploadUrl: djaodjinSettings.urls.pages.api_themes,
             });
-        });
+
+            $("#reset-theme").click(function(event) {
+                $.ajax({
+                    method: "DELETE",
+                    url: djaodjinSettings.urls.pages.api_themes,
+                    datatype: "json",
+                    contentType: "application/json; charset=utf-8",
+                    success: function(resp) {
+                        showMessages([gettext("reset to default theme")], "success");
+                    },
+                    error: function(resp) {
+                        showErrorMessages(resp);
+                    }
+                });
+            });
+        }
     });
 
 })(jQuery);
